repair landingpage

+++ database/migrations/2024_02_14_050101_create_programs_table.php
@@ -40,6 +40,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('programs');
+        if (Schema::hasTable('programs')) {
+            Schema::dropIfExists('programs');
+        }
     }
 };
